CI,commitlint(tests): upgrade to latest commitlint
Convert commitlint plugins to ES module so that it can be
loaded with "module": "node16" setting that's needed to load
commitlint, which is ES module now.

Run tsc for tests separately and have separate tsconfig.json
for tests directory as tests use CommonJS modules which are
incompatible with ES modules.